Add support function files
